refactor: fold resources_test.go into reconciler_test.go
Per Nir's feedback — consistent with the resources.go fold done earlier.
Test functions moved as-is, no logic changes.
