+++ pkg/controller/externalmodel/reconciler_test.go
@@ -320,3 +320,82 @@ func TestReconcile_TwoModelsOneProvider(t *testing.T) {
 	assert.Equal(t, "gpt4", hr1.OwnerReferences[0].Name)
 	assert.Equal(t, "gpt35", hr2.OwnerReferences[0].Name)
 }
+
+// --- resource builder tests (folded from resources_test.go) ---
+
+func TestCommonLabels(t *testing.T) {
+	labels := commonLabels("gpt4")
+	assert.Equal(t, managedByValue, labels[ctrlcommon.LabelManagedBy])
+	assert.Equal(t, "gpt4", labels[labelExternalModel])
+	assert.Len(t, labels, 2)
+}
+
+func TestBuildHTTPRoute(t *testing.T) {
+	hr := buildHTTPRoute(
+		"api.openai.com", "my-openai",
+		"gpt4", "gpt-4o",
+		"models", 443,
+		"default-gateway", "openshift-ingress", "300s",
+		commonLabels("gpt4"),
+	)
+
+	assert.Equal(t, "gpt4", hr.Name)
+	assert.Equal(t, "models", hr.Namespace)
+	assert.Equal(t, managedByValue, hr.Labels[ctrlcommon.LabelManagedBy])
+
+	// Parent gateway ref
+	require.Len(t, hr.Spec.ParentRefs, 1)
+	assert.Equal(t, "default-gateway", string(hr.Spec.ParentRefs[0].Name))
+	assert.Equal(t, "openshift-ingress", string(*hr.Spec.ParentRefs[0].Namespace))
+
+	// Must have 2 rules: path-based and header-based
+	require.Len(t, hr.Spec.Rules, 2)
+
+	// Rule 1: path-based match with namespace prefix
+	rule1 := hr.Spec.Rules[0]
+	assert.Equal(t, "/models/gpt4", *rule1.Matches[0].Path.Value)
+
+	// Rule 2: header-based match uses targetModel
+	rule2 := hr.Spec.Rules[1]
+	assert.Equal(t, "X-Gateway-Model-Name", string(rule2.Matches[0].Headers[0].Name))
+	assert.Equal(t, "gpt-4o", rule2.Matches[0].Headers[0].Value)
+
+	// Backend ref points to the PROVIDER's Service, not the model
+	for i, rule := range hr.Spec.Rules {
+		require.Len(t, rule.BackendRefs, 1, "rule %d", i)
+		assert.Equal(t, "my-openai", string(rule.BackendRefs[0].Name),
+			"rule %d: backend should be the provider's Service", i)
+	}
+
+	// Host header filter for TLS SNI uses provider endpoint
+	for i, rule := range hr.Spec.Rules {
+		require.Len(t, rule.Filters, 1, "rule %d", i)
+		assert.Equal(t, gatewayapiv1.HTTPRouteFilterRequestHeaderModifier, rule.Filters[0].Type)
+		assert.Equal(t, "Host", string(rule.Filters[0].RequestHeaderModifier.Set[0].Name))
+		assert.Equal(t, "api.openai.com", rule.Filters[0].RequestHeaderModifier.Set[0].Value)
+	}
+}
+
+func TestBuildHTTPRoute_TargetModelDiffersFromName(t *testing.T) {
+	hr := buildHTTPRoute(
+		"bedrock.us-east-1.amazonaws.com", "my-bedrock",
+		"claude", "anthropic.claude-3-opus",
+		"models", 443,
+		"my-gateway", "gateway-ns", "300s",
+		commonLabels("claude"),
+	)
+
+	// Name and path use ExternalModel name
+	assert.Equal(t, "claude", hr.Name)
+	assert.Equal(t, "/models/claude", *hr.Spec.Rules[0].Matches[0].Path.Value)
+
+	// Header match uses targetModel (provider-side name)
+	assert.Equal(t, "anthropic.claude-3-opus", hr.Spec.Rules[1].Matches[0].Headers[0].Value)
+
+	// Backend points to provider Service
+	assert.Equal(t, "my-bedrock", string(hr.Spec.Rules[0].BackendRefs[0].Name))
+
+	// Host header uses provider endpoint
+	assert.Equal(t, "bedrock.us-east-1.amazonaws.com",
+		hr.Spec.Rules[0].Filters[0].RequestHeaderModifier.Set[0].Value)
+}
